requirements.txt
PyQt6>=6.7.0
Pillow>=9.0.0